Message_Channel/InvalidMessage/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
kafka-python==2.0.4
MarkupSafe==3.0.2
Werkzeug==3.1.3